+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 - `VvForm` bug with emit update on zod parsed result;
 - Experimental components slots types;
 - `VvFormTemplate` component;
-- typescript improvements.
+- Typescript improvements.
 
 ### Added